+++ public/ndla-favicon.png
