--- server/developmentAssets.json
@@ -1,9 +1,0 @@
-{
-  "ndla-favicon.png": "ndla-favicon.png",
-  "main.css": "main.css",
-  "main.js": "main.js",
-  "manifest.js": "manifest.js",
-  "vendor.js": "vendor.js",
-  "embed.js": "embed.js",
-  "mathjaxConfig.js": "mathjaxConfig.js"
-}
