--- server/helpers/enableDevMiddleware.js
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) 2016-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-const webpack = require('webpack');
-const webpackDevMiddleware = require('webpack-dev-middleware');
-const webpackHotMiddleware = require('webpack-hot-middleware');
-
-export default function enableDevMiddleWare(app) {
-  const webpackConfig = require('../../webpack.config.dev'); // eslint-disable-line global-require
-  const compiler = webpack(webpackConfig);
-  app.use(
-    webpackDevMiddleware(compiler, {
-      noInfo: true,
-      stats: {
-        colors: true,
-      },
-      publicPath: webpackConfig.output.publicPath,
-    }),
-  );
-  app.use(webpackHotMiddleware(compiler, {}));
-}
