+++ src/containers/Resources/resourceHelpers.js
@@ -10,7 +10,7 @@ import config from '../../config';
 import { toArticle } from '../../routeHelpers';
 
 const LEARNING_PATH_DOMAIN =
-  __SERVER__ || process.env.NODE_ENV === 'unittest'
+  process.env.BUILD_TARGET === 'server' || process.env.NODE_ENV === 'unittest'
     ? config.learningPathDomain
     : window.DATA.config.learningPathDomain;
 export const URN_ARTICLE = 'urn:article:';
