+++ src/util/handleError.js
@@ -9,11 +9,17 @@
 import ErrorReporter from 'ndla-error-reporter';
 
 export default error => {
-  if (process.env.NODE_ENV === 'production' && __CLIENT__) {
+  if (
+    process.env.NODE_ENV === 'production' &&
+    process.env.BUILD_TARGET === 'client'
+  ) {
     ErrorReporter.getInstance().captureError(error);
-  } else if (__SERVER__) {
-    const log = require('./logger'); // eslint-disable-line global-require
-    log.error(error);
+    // } else if (
+    //   process.env.NODE_ENV === 'production' &&
+    //   process.env.BUILD_TARGET === 'server'
+    // ) {
+    //   const log = require('./logger'); // eslint-disable-line global-require
+    //   log.error(error);
   } else {
     console.error(error); // eslint-disable-line no-console
   }
